Add exercise options to the modal, according to the corresponding json file that the user chose.
Changed some css to give a scroll function in case
the options are too many, changed javascript to
read json exercises in try function, did the
unsynchronous function because of the fecth, but
did the fetch in a synchronous way with the await.
Exercises only have the name for now and cannot
choose any.

[Ticket: X]